feat: refactor model types and enhance component imports
